requirements.txt
# Reddit API
praw==7.7.1

# Google Sheets integration
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Data processing and utilities
requests==2.31.0
python-dotenv==1.0.0

# Optional: For enhanced text processing
nltk==3.8.1
textblob==0.17.1

# Development and testing
pytest==7.4.3
python-dateutil==2.8.2